Renaming *_axis to *_axes
