+++ src/tensor_ops/mod.rs
@@ -119,21 +119,21 @@ mod impl_clamp;
 mod impl_div;
 mod impl_dropout;
 mod impl_mask;
-mod impl_max_axis;
+mod impl_max_axes;
 mod impl_maximum;
 mod impl_mean;
-mod impl_mean_axis;
-mod impl_min_axis;
+mod impl_mean_axes;
+mod impl_min_axes;
 mod impl_minimum;
 mod impl_mul;
 mod impl_nans;
-mod impl_normalize_axis;
+mod impl_normalize_axes;
 mod impl_pow;
 mod impl_softmax;
-mod impl_std_axis;
+mod impl_std_axes;
 mod impl_sub;
 mod impl_sum;
-mod impl_sum_axis;
+mod impl_sum_axes;
 mod map;
 mod matmul;
 mod permute;
@@ -148,21 +148,21 @@ pub use impl_clamp::*;
 pub use impl_div::*;
 pub use impl_dropout::*;
 pub use impl_mask::*;
-pub use impl_max_axis::*;
+pub use impl_max_axes::*;
 pub use impl_maximum::*;
 pub use impl_mean::*;
-pub use impl_mean_axis::*;
-pub use impl_min_axis::*;
+pub use impl_mean_axes::*;
+pub use impl_min_axes::*;
 pub use impl_minimum::*;
 pub use impl_mul::*;
 pub use impl_nans::*;
-pub use impl_normalize_axis::*;
+pub use impl_normalize_axes::*;
 pub use impl_pow::*;
 pub use impl_softmax::*;
-pub use impl_std_axis::*;
+pub use impl_std_axes::*;
 pub use impl_sub::*;
 pub use impl_sum::*;
-pub use impl_sum_axis::*;
+pub use impl_sum_axes::*;
 pub use map::*;
 pub use matmul::*;
 pub use permute::{Permute2D, Permute3D, Permute4D};
